# pkg/openapi/bonus/v1/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package bonus_v1

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/user/"

			if l := len("/api/user/"); len(elem) >= l && elem[0:l] == "/api/user/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'b': // Prefix: "balance"

				if l := len("balance"); len(elem) >= l && elem[0:l] == "balance" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleGetUserBalanceRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET")
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/withdraw"

					if l := len("/withdraw"); len(elem) >= l && elem[0:l] == "/withdraw" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleBalanceWithdrawalRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "POST")
						}

						return
					}
				}

			case 'l': // Prefix: "login"

				if l := len("login"); len(elem) >= l && elem[0:l] == "login" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleLoginUserRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "POST")
					}

					return
				}

			case 'o': // Prefix: "orders"

				if l := len("orders"); len(elem) >= l && elem[0:l] == "orders" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetOrdersNumberListRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleOrderNumberLoadRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET,POST")
					}

					return
				}

			case 'r': // Prefix: "register"

				if l := len("register"); len(elem) >= l && elem[0:l] == "register" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleRegisterUserRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "POST")
					}

					return
				}

			case 'w': // Prefix: "withdrawals"

				if l := len("withdrawals"); len(elem) >= l && elem[0:l] == "withdrawals" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetWithdrawalsRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET")
					}

					return
				}

			}
		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [0]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/api/user/"

			if l := len("/api/user/"); len(elem) >= l && elem[0:l] == "/api/user/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'b': // Prefix: "balance"

				if l := len("balance"); len(elem) >= l && elem[0:l] == "balance" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = GetUserBalanceOperation
						r.summary = "Получение текущего баланса пользователя"
						r.operationID = "GetUserBalance"
						r.operationGroup = ""
						r.pathPattern = "/api/user/balance"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/withdraw"

					if l := len("/withdraw"); len(elem) >= l && elem[0:l] == "/withdraw" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = BalanceWithdrawalOperation
							r.summary = "Запрос на списание средств"
							r.operationID = "BalanceWithdrawal"
							r.operationGroup = ""
							r.pathPattern = "/api/user/balance/withdraw"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
				}

			case 'l': // Prefix: "login"

				if l := len("login"); len(elem) >= l && elem[0:l] == "login" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = LoginUserOperation
						r.summary = "Аутентификация пользователя"
						r.operationID = "LoginUser"
						r.operationGroup = ""
						r.pathPattern = "/api/user/login"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'o': // Prefix: "orders"

				if l := len("orders"); len(elem) >= l && elem[0:l] == "orders" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetOrdersNumberListOperation
						r.summary = "Получение списка загруженных номеров заказов"
						r.operationID = "GetOrdersNumberList"
						r.operationGroup = ""
						r.pathPattern = "/api/user/orders"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = OrderNumberLoadOperation
						r.summary = "Загрузка номера заказа"
						r.operationID = "OrderNumberLoad"
						r.operationGroup = ""
						r.pathPattern = "/api/user/orders"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'r': // Prefix: "register"

				if l := len("register"); len(elem) >= l && elem[0:l] == "register" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = RegisterUserOperation
						r.summary = "Регистрация пользователя"
						r.operationID = "RegisterUser"
						r.operationGroup = ""
						r.pathPattern = "/api/user/register"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			case 'w': // Prefix: "withdrawals"

				if l := len("withdrawals"); len(elem) >= l && elem[0:l] == "withdrawals" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetWithdrawalsOperation
						r.summary = "Получение информации о выводе средств"
						r.operationID = "GetWithdrawals"
						r.operationGroup = ""
						r.pathPattern = "/api/user/withdrawals"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

			}
		}
	}
	return r, false
}